dashboard_control: carve wire.rs (pure move)
The control-channel wire loop: per-peer driver task, outbound senders,
output draining, frame-text assembly and chunking, queued-frame drain;
five wire-format tests ride along. The fs_scope_grant_tests mod stays
in mod.rs beside the grant impl it exercises.
