+++ src/bin/caller/dashboard_control/mod.rs
@@ -39,6 +39,9 @@ use tokio::net::UdpSocket;
 use tokio::sync::{mpsc, Mutex};
 use tokio_util::sync::CancellationToken;
 
+mod wire;
+pub(crate) use wire::*;
+
 const CONTROL_CHANNEL_LABEL: &str = "intendant-dashboard-control";
 const CONTROL_PROTOCOL_VERSION: u32 = 1;
 const CONTROL_SIGNATURE_CONTEXT: &str = "intendant-dashboard-control-v1";
@@ -1349,627 +1352,6 @@ impl OutboundControlQueue {
     }
 }
 
-async fn control_driver<I: rtc::interceptor::Interceptor + Send + Sync + 'static>(
-    mut rtc: RTCPeerConnection<I>,
-    sockets: Vec<Arc<UdpSocket>>,
-    mut tcp_conn_rx: Option<mpsc::Receiver<crate::display::webrtc::AcceptedTcpConnection>>,
-    tcp_advertised: Option<SocketAddr>,
-    _tcp_registration: Option<crate::display::webrtc::PeerRegistration>,
-    mut runtime: ControlRuntime,
-    mut event_rx: tokio::sync::broadcast::Receiver<String>,
-    mut command_rx: mpsc::Receiver<ControlCommand>,
-    shutdown: CancellationToken,
-) {
-    let mut sockets_by_addr: HashMap<SocketAddr, Arc<UdpSocket>> = HashMap::new();
-    let (inbound_tx, mut inbound_rx) = mpsc::channel::<InboundPacket>(64);
-    let mut forwarder_handles = Vec::new();
-    for sock in sockets {
-        let local = match sock.local_addr() {
-            Ok(local) => local,
-            Err(_) => continue,
-        };
-        sockets_by_addr.insert(local, Arc::clone(&sock));
-        let tx = inbound_tx.clone();
-        let shutdown = shutdown.clone();
-        forwarder_handles.push(tokio::spawn(async move {
-            let mut buf = vec![0u8; UDP_BUF_LEN];
-            loop {
-                tokio::select! {
-                    _ = shutdown.cancelled() => break,
-                    recv = sock.recv_from(&mut buf) => match recv {
-                        Ok((n, source)) => {
-                            let pkt = InboundPacket {
-                                proto: TransportProtocol::UDP,
-                                source,
-                                destination: local,
-                                bytes: buf[..n].to_vec(),
-                                received_at: Instant::now(),
-                            };
-                            if tx.send(pkt).await.is_err() {
-                                break;
-                            }
-                        }
-                        Err(e) => {
-                            eprintln!("[dashboard/control] UDP recv failed on {local}: {e}");
-                            break;
-                        }
-                    }
-                }
-            }
-        }));
-    }
-    let mut tcp_senders: HashMap<SocketAddr, TcpFrameSender> = HashMap::new();
-    let mut channels: HashMap<String, rtc::data_channel::RTCDataChannelId> = HashMap::new();
-    let (task_tx, mut task_rx) = mpsc::channel::<ControlTaskResponse>(64);
-    let mut pending_requests: HashMap<String, CancellationToken> = HashMap::new();
-    let mut outbound_queue = OutboundControlQueue::new();
-    let mut inbound_uploads: HashMap<String, InboundUploadState> = HashMap::new();
-    let (terminal_events_tx, mut terminal_events_rx) =
-        mpsc::unbounded_channel::<serde_json::Value>();
-    runtime.control_frames_tx = Some(terminal_events_tx.clone());
-    let mut terminal_forwarders: HashMap<(String, String), tokio::task::JoinHandle<()>> =
-        HashMap::new();
-    let mut display_authority_rx = runtime
-        .display_authority
-        .as_ref()
-        .map(DashboardDisplayAuthorityBridge::subscribe);
-    let mut drop_stats = TransmitDropStats::default();
-
-    loop {
-        let timeout_at = match drain_control_outputs(
-            &mut rtc,
-            &sockets_by_addr,
-            &mut tcp_senders,
-            &mut drop_stats,
-            &mut channels,
-            &mut runtime,
-            &task_tx,
-            &mut pending_requests,
-            &mut outbound_queue,
-            &mut inbound_uploads,
-            &terminal_events_tx,
-            &mut terminal_forwarders,
-        )
-        .await
-        {
-            Ok(timeout_at) => timeout_at,
-            Err(()) => {
-                shutdown.cancel();
-                break;
-            }
-        };
-        let timeout_dur = timeout_at
-            .saturating_duration_since(Instant::now())
-            .max(Duration::from_micros(1));
-
-        tokio::select! {
-            _ = shutdown.cancelled() => break,
-            Some(pkt) = inbound_rx.recv() => {
-                let input = TaggedBytesMut {
-                    now: pkt.received_at,
-                    transport: TransportContext {
-                        local_addr: pkt.destination,
-                        peer_addr: pkt.source,
-                        transport_protocol: pkt.proto,
-                        ecn: None,
-                    },
-                    message: BytesMut::from(pkt.bytes.as_slice()),
-                };
-                if let Err(e) = rtc.handle_read(input) {
-                    eprintln!("[dashboard/control] handle_read failed: {e:?}");
-                    shutdown.cancel();
-                    break;
-                }
-            }
-            Some(accepted) = async {
-                match tcp_conn_rx.as_mut() {
-                    Some(rx) => rx.recv().await,
-                    None => std::future::pending().await,
-                }
-            } => {
-                let crate::display::webrtc::AcceptedTcpConnection {
-                    remote_addr,
-                    local_addr: real_local,
-                    first_frame,
-                    stream,
-                } = accepted;
-                let Some(fake_local) = tcp_advertised else {
-                    eprintln!(
-                        "[dashboard/control] TCP connection from {remote_addr} but no advertised local configured, dropping"
-                    );
-                    continue;
-                };
-                eprintln!(
-                    "[dashboard/control] ICE-TCP connection from {remote_addr} -> {real_local} (rtc sees {fake_local})"
-                );
-                let (read_half, write_half) = stream.into_split();
-                let (tcp_out_tx, mut tcp_out_rx) = mpsc::channel::<Vec<u8>>(TCP_OUT_QUEUE);
-                tcp_senders.insert(remote_addr, tcp_out_tx);
-
-                let writer_shutdown = shutdown.clone();
-                tokio::spawn(async move {
-                    let mut write_half = write_half;
-                    loop {
-                        tokio::select! {
-                            biased;
-                            _ = writer_shutdown.cancelled() => break,
-                            frame = tcp_out_rx.recv() => match frame {
-                                Some(contents) => {
-                                    if let Err(e) =
-                                        crate::display::webrtc::write_rfc4571_frame(&mut write_half, &contents).await
-                                    {
-                                        eprintln!(
-                                            "[dashboard/control] ICE-TCP writer for {remote_addr} failed, tearing down connection: {e}"
-                                        );
-                                        writer_shutdown.cancel();
-                                        break;
-                                    }
-                                }
-                                None => break,
-                            }
-                        }
-                    }
-                    let _ = tokio::io::AsyncWriteExt::shutdown(&mut write_half).await;
-                });
-
-                let reader_tx = inbound_tx.clone();
-                let reader_shutdown = shutdown.clone();
-                tokio::spawn(async move {
-                    let mut read_half = read_half;
-                    loop {
-                        tokio::select! {
-                            _ = reader_shutdown.cancelled() => break,
-                            frame = crate::display::webrtc::read_rfc4571_frame(&mut read_half) => match frame {
-                                Ok(bytes) => {
-                                    let pkt = InboundPacket {
-                                        proto: TransportProtocol::TCP,
-                                        source: remote_addr,
-                                        destination: fake_local,
-                                        bytes,
-                                        received_at: Instant::now(),
-                                    };
-                                    if reader_tx.send(pkt).await.is_err() {
-                                        break;
-                                    }
-                                }
-                                Err(e) => {
-                                    eprintln!(
-                                        "[dashboard/control] ICE-TCP reader for {remote_addr} exiting: {e}"
-                                    );
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                });
-
-                let input = TaggedBytesMut {
-                    now: Instant::now(),
-                    transport: TransportContext {
-                        local_addr: fake_local,
-                        peer_addr: remote_addr,
-                        transport_protocol: TransportProtocol::TCP,
-                        ecn: None,
-                    },
-                    message: BytesMut::from(first_frame.as_slice()),
-                };
-                if let Err(e) = rtc.handle_read(input) {
-                    eprintln!("[dashboard/control] handle_read(first TCP frame) failed: {e:?}");
-                    shutdown.cancel();
-                    break;
-                }
-            }
-            Some(cmd) = command_rx.recv() => {
-                match cmd {
-                    ControlCommand::AddIceCandidate(candidate) => {
-                        let init = RTCIceCandidateInit {
-                            candidate,
-                            sdp_mid: None,
-                            sdp_mline_index: None,
-                            username_fragment: None,
-                            url: None,
-                        };
-                        if let Err(e) = rtc.add_remote_candidate(init) {
-                            eprintln!("[dashboard/control] parse remote candidate failed: {e}");
-                        }
-                    }
-                }
-                let _ = rtc.handle_timeout(Instant::now());
-            }
-            Some(task_response) = task_rx.recv() => {
-                if pending_requests.contains_key(&task_response.id) {
-                    let task_id = task_response.id.clone();
-                    let done = task_response.done;
-                    send_control_task_response(
-                        &mut rtc,
-                        &channels,
-                        &mut outbound_queue,
-                        runtime.response_credit_enabled,
-                        task_response,
-                    );
-                    if done {
-                        pending_requests.remove(&task_id);
-                    }
-                }
-                let _ = rtc.handle_timeout(Instant::now());
-            }
-            event = event_rx.recv(), if runtime.events_subscribed => {
-                match event {
-                    Ok(line) => {
-                        runtime.events_sent = runtime.events_sent.saturating_add(1);
-                        let payload = serde_json::from_str::<serde_json::Value>(&line)
-                            .unwrap_or_else(|_| serde_json::json!({"raw": line}));
-                        let frame = serde_json::json!({
-                            "t": "event",
-                            "seq": runtime.events_sent,
-                            "payload": payload,
-                        });
-                        send_control_text(&mut rtc, &channels, frame.to_string());
-                    }
-                    Err(tokio::sync::broadcast::error::RecvError::Lagged(skipped)) => {
-                        let frame = serde_json::json!({
-                            "t": "event_gap",
-                            "skipped": skipped,
-                        });
-                        send_control_text(&mut rtc, &channels, frame.to_string());
-                    }
-                    Err(tokio::sync::broadcast::error::RecvError::Closed) => {
-                        runtime.events_subscribed = false;
-                    }
-                }
-                let _ = rtc.handle_timeout(Instant::now());
-            }
-            Some(frame) = terminal_events_rx.recv() => {
-                send_control_text(&mut rtc, &channels, frame.to_string());
-                let _ = rtc.handle_timeout(Instant::now());
-            }
-            authority = async {
-                match display_authority_rx.as_mut() {
-                    Some(rx) => Some(rx.recv().await),
-                    None => std::future::pending::<Option<Result<u32, tokio::sync::broadcast::error::RecvError>>>().await,
-                }
-            }, if runtime.events_subscribed && display_authority_rx.is_some() => {
-                match authority {
-                    Some(Ok(display_id)) => {
-                        send_display_authority_event(&mut rtc, &channels, &mut runtime, display_id);
-                    }
-                    Some(Err(tokio::sync::broadcast::error::RecvError::Lagged(_))) => {
-                        let snapshots = display_authority_snapshot_frames(&runtime).await;
-                        for frame in snapshots {
-                            send_event_payload(&mut rtc, &channels, &mut runtime, frame);
-                        }
-                    }
-                    Some(Err(tokio::sync::broadcast::error::RecvError::Closed)) | None => {
-                        display_authority_rx = runtime
-                            .display_authority
-                            .as_ref()
-                            .map(DashboardDisplayAuthorityBridge::subscribe);
-                    }
-                }
-                let _ = rtc.handle_timeout(Instant::now());
-            }
-            _ = tokio::time::sleep(timeout_dur) => {
-                if let Err(e) = rtc.handle_timeout(Instant::now()) {
-                    eprintln!("[dashboard/control] handle_timeout failed: {e:?}");
-                    shutdown.cancel();
-                    break;
-                }
-            }
-        }
-    }
-
-    for (_, token) in pending_requests {
-        token.cancel();
-    }
-    for (_, handle) in terminal_forwarders {
-        handle.abort();
-        let _ = handle.await;
-    }
-    // Egress relays die with their session: no more frames can arrive,
-    // so drop the registration and fail any in-flight relayed requests.
-    crate::credential_egress::unregister_session(&runtime.session_id);
-    if let Some(bridge) = &runtime.presence {
-        bridge.cleanup(runtime.session_id.clone()).await;
-    }
-    for handle in forwarder_handles {
-        let _ = handle.await;
-    }
-}
-
-fn send_event_payload<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    channels: &HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    runtime: &mut ControlRuntime,
-    payload: serde_json::Value,
-) {
-    runtime.events_sent = runtime.events_sent.saturating_add(1);
-    let frame = serde_json::json!({
-        "t": "event",
-        "seq": runtime.events_sent,
-        "payload": payload,
-    });
-    send_control_text(rtc, channels, frame.to_string());
-}
-
-fn send_display_authority_event<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    channels: &HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    runtime: &mut ControlRuntime,
-    display_id: u32,
-) {
-    let Some(bridge) = runtime.display_authority.as_ref() else {
-        return;
-    };
-    if let Some(frame) = bridge.state_frame(&runtime.session_id, display_id) {
-        send_event_payload(rtc, channels, runtime, frame);
-    }
-}
-
-async fn drain_control_outputs<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    sockets_by_addr: &HashMap<SocketAddr, Arc<UdpSocket>>,
-    tcp_senders: &mut HashMap<SocketAddr, TcpFrameSender>,
-    drop_stats: &mut TransmitDropStats,
-    channels: &mut HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    runtime: &mut ControlRuntime,
-    task_tx: &mpsc::Sender<ControlTaskResponse>,
-    pending_requests: &mut HashMap<String, CancellationToken>,
-    outbound_queue: &mut OutboundControlQueue,
-    inbound_uploads: &mut HashMap<String, InboundUploadState>,
-    terminal_events_tx: &mpsc::UnboundedSender<serde_json::Value>,
-    terminal_forwarders: &mut HashMap<(String, String), tokio::task::JoinHandle<()>>,
-) -> Result<Instant, ()> {
-    while let Some(t) = rtc.poll_write() {
-        // Route by connection before trusting the engine's protocol stamp:
-        // rtc 0.9 marks DTLS and SCTP transmits `TransportProtocol::UDP`
-        // even when the selected pair is TCP ("TransportProtocol doesn't
-        // matter" — rtc/src/peer_connection/transport/dtls/mod.rs), so a
-        // peer that reached us over ICE-TCP must be matched by its tuple,
-        // not by the stamp, or every post-ICE packet misses the stream and
-        // DTLS times out.
-        if let Some(sender) = tcp_senders.get(&t.transport.peer_addr) {
-            let contents = t.message.to_vec();
-            match sender.try_send(contents) {
-                Ok(()) => {}
-                Err(mpsc::error::TrySendError::Full(_)) => {}
-                Err(mpsc::error::TrySendError::Closed(_)) => {
-                    tcp_senders.remove(&t.transport.peer_addr);
-                }
-            }
-            continue;
-        }
-        if t.transport.transport_protocol == TransportProtocol::TCP {
-            // TCP-stamped transmit with no live stream for the tuple: the
-            // connection is gone and there is nothing to write to.
-            drop_stats.tcp_without_stream += 1;
-            continue;
-        }
-        if t.transport.local_addr.is_ipv4() != t.transport.peer_addr.is_ipv4() {
-            drop_stats.cross_family += 1;
-            continue;
-        }
-        if t.transport.local_addr.ip().is_loopback() != t.transport.peer_addr.ip().is_loopback() {
-            drop_stats.loopback_mismatch += 1;
-            continue;
-        }
-        let Some(sock) = sockets_by_addr.get(&t.transport.local_addr) else {
-            drop_stats.unknown_udp_source += 1;
-            eprintln!(
-                "[dashboard/control] UDP transmit from unknown source {}, dropping",
-                t.transport.local_addr
-            );
-            continue;
-        };
-        if let Err(e) = sock.send_to(&t.message, t.transport.peer_addr).await {
-            eprintln!(
-                "[dashboard/control] udp send {} -> {} failed: {e}",
-                t.transport.local_addr, t.transport.peer_addr
-            );
-        }
-    }
-
-    while let Some(message) = rtc.poll_read() {
-        let RTCMessage::DataChannelMessage(cid, msg) = message else {
-            continue;
-        };
-        let label = channels
-            .iter()
-            .find_map(|(label, id)| (*id == cid).then(|| label.clone()));
-        if label.as_deref() != Some(CONTROL_CHANNEL_LABEL) {
-            continue;
-        }
-        let Ok(text) = std::str::from_utf8(&msg.data) else {
-            continue;
-        };
-        if let Some(response) = control_frame_response(
-            text,
-            runtime,
-            task_tx,
-            pending_requests,
-            outbound_queue,
-            inbound_uploads,
-            terminal_events_tx,
-            terminal_forwarders,
-        ) {
-            send_control_frame(
-                rtc,
-                channels,
-                outbound_queue,
-                runtime.response_credit_enabled,
-                response,
-            );
-        }
-    }
-
-    while let Some(event) = rtc.poll_event() {
-        match event {
-            RTCPeerConnectionEvent::OnDataChannel(RTCDataChannelEvent::OnOpen(cid)) => {
-                let label = rtc
-                    .data_channel(cid)
-                    .map(|channel| channel.label().to_string())
-                    .unwrap_or_else(|| format!("channel-{cid}"));
-                eprintln!("[dashboard/control] data channel open: {label}");
-                channels.insert(label, cid);
-            }
-            RTCPeerConnectionEvent::OnDataChannel(RTCDataChannelEvent::OnClose(cid)) => {
-                channels.retain(|_, id| *id != cid);
-            }
-            RTCPeerConnectionEvent::OnConnectionStateChangeEvent(state) => {
-                eprintln!("[dashboard/control] connection: {state:?}");
-                if matches!(
-                    state,
-                    rtc::peer_connection::state::RTCPeerConnectionState::Failed
-                        | rtc::peer_connection::state::RTCPeerConnectionState::Closed
-                ) {
-                    if drop_stats.any() {
-                        eprintln!(
-                            "[dashboard/control] transmit drops this session: {} cross-family, {} loopback-mismatch, {} unknown-udp-source, {} tcp-without-stream",
-                            drop_stats.cross_family,
-                            drop_stats.loopback_mismatch,
-                            drop_stats.unknown_udp_source,
-                            drop_stats.tcp_without_stream
-                        );
-                    }
-                    return Err(());
-                }
-            }
-            RTCPeerConnectionEvent::OnIceConnectionStateChangeEvent(state) => {
-                eprintln!("[dashboard/control] ICE: {state:?}");
-            }
-            _ => {}
-        }
-    }
-
-    drain_queued_control_frames(rtc, channels, outbound_queue);
-
-    Ok(rtc
-        .poll_timeout()
-        .unwrap_or_else(|| Instant::now() + Duration::from_secs(86_400)))
-}
-
-fn send_control_text<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    channels: &HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    text: String,
-) {
-    let Some(cid) = channels.get(CONTROL_CHANNEL_LABEL).copied() else {
-        return;
-    };
-    if let Some(mut channel) = rtc.data_channel(cid) {
-        if let Err(e) = channel.send_text(text) {
-            eprintln!("[dashboard/control] data channel write failed: {e:?}");
-        }
-    }
-}
-
-fn send_control_task_response<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    channels: &HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    outbound_queue: &mut OutboundControlQueue,
-    response_credit_enabled: bool,
-    response: ControlTaskResponse,
-) {
-    if let Some(byte_stream) = response.byte_stream {
-        send_control_byte_stream(
-            rtc,
-            channels,
-            outbound_queue,
-            response_credit_enabled,
-            byte_stream,
-        );
-    } else {
-        send_control_frame(
-            rtc,
-            channels,
-            outbound_queue,
-            response_credit_enabled,
-            response.frame,
-        );
-    }
-}
-
-fn send_control_frame<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    channels: &HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    outbound_queue: &mut OutboundControlQueue,
-    response_credit_enabled: bool,
-    frame: serde_json::Value,
-) {
-    let request_id = frame
-        .get("id")
-        .and_then(|v| v.as_str())
-        .unwrap_or("")
-        .to_string();
-    match control_frame_text_parts(
-        frame,
-        CONTROL_RESPONSE_CHUNK_THRESHOLD_BYTES,
-        CONTROL_RESPONSE_CHUNK_BYTES,
-    ) {
-        ControlFrameTexts::Immediate(frames) => {
-            for text in frames {
-                if response_credit_enabled && !outbound_queue.is_empty() && !request_id.is_empty() {
-                    outbound_queue.enqueue_immediate(request_id.clone(), text);
-                } else {
-                    send_control_text(rtc, channels, text);
-                }
-            }
-            drain_queued_control_frames(rtc, channels, outbound_queue);
-        }
-        ControlFrameTexts::Chunked {
-            request_id,
-            chunk_id,
-            start,
-            chunks,
-            end,
-        } => {
-            if response_credit_enabled {
-                outbound_queue.enqueue_chunked(request_id, chunk_id, start, chunks, end);
-                drain_queued_control_frames(rtc, channels, outbound_queue);
-            } else {
-                send_control_text(rtc, channels, start);
-                for text in chunks {
-                    send_control_text(rtc, channels, text);
-                }
-                send_control_text(rtc, channels, end);
-            }
-        }
-    }
-}
-
-fn send_control_byte_stream<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    channels: &HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    outbound_queue: &mut OutboundControlQueue,
-    response_credit_enabled: bool,
-    byte_stream: ControlByteStream,
-) {
-    match byte_stream_frame_text_parts(byte_stream, CONTROL_BYTE_STREAM_CHUNK_BYTES) {
-        ControlFrameTexts::Immediate(frames) => {
-            for text in frames {
-                send_control_text(rtc, channels, text);
-            }
-        }
-        ControlFrameTexts::Chunked {
-            request_id,
-            chunk_id,
-            start,
-            chunks,
-            end,
-        } => {
-            if response_credit_enabled {
-                outbound_queue.enqueue_chunked(request_id, chunk_id, start, chunks, end);
-                drain_queued_control_frames(rtc, channels, outbound_queue);
-            } else {
-                send_control_text(rtc, channels, start);
-                for text in chunks {
-                    send_control_text(rtc, channels, text);
-                }
-                send_control_text(rtc, channels, end);
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod fs_scope_grant_tests {
     use super::*;
@@ -2026,242 +1408,6 @@ mod fs_scope_grant_tests {
         .filesystem()
         .is_none());
     }
-}
-
-#[cfg(test)]
-fn control_frame_texts(frame: serde_json::Value) -> Vec<String> {
-    match control_frame_text_parts(
-        frame,
-        CONTROL_RESPONSE_CHUNK_THRESHOLD_BYTES,
-        CONTROL_RESPONSE_CHUNK_BYTES,
-    ) {
-        ControlFrameTexts::Immediate(frames) => frames,
-        ControlFrameTexts::Chunked {
-            start, chunks, end, ..
-        } => {
-            let mut frames = Vec::with_capacity(chunks.len() + 2);
-            frames.push(start);
-            frames.extend(chunks);
-            frames.push(end);
-            frames
-        }
-    }
-}
-
-fn byte_stream_frame_text_parts(
-    byte_stream: ControlByteStream,
-    chunk_bytes: usize,
-) -> ControlFrameTexts {
-    let request_id = byte_stream.id;
-    let chunk_id = byte_stream.stream_id;
-    if request_id.is_empty() || chunk_id.is_empty() || chunk_bytes == 0 {
-        return ControlFrameTexts::Immediate(Vec::new());
-    }
-
-    let total_bytes = byte_stream.bytes.len();
-    let chunk_count = total_bytes.div_ceil(chunk_bytes);
-    let start = serde_json::json!({
-        "t": "byte_stream_start",
-        "id": request_id,
-        "stream_id": chunk_id,
-        "encoding": "base64",
-        "content_type": byte_stream.content_type,
-        "filename": byte_stream.filename,
-        "total_bytes": total_bytes,
-        "chunks": chunk_count,
-    })
-    .to_string();
-    let mut chunks = Vec::with_capacity(chunk_count);
-    for (seq, chunk) in byte_stream.bytes.chunks(chunk_bytes).enumerate() {
-        chunks.push(
-            serde_json::json!({
-                "t": "byte_stream_chunk",
-                "id": request_id,
-                "stream_id": chunk_id,
-                "seq": seq,
-                "data": base64::engine::general_purpose::STANDARD.encode(chunk),
-            })
-            .to_string(),
-        );
-    }
-    let end = serde_json::json!({
-        "t": "byte_stream_end",
-        "id": request_id,
-        "stream_id": chunk_id,
-        "ok": true,
-        "chunks": chunk_count,
-        "result": byte_stream.result,
-    })
-    .to_string();
-    ControlFrameTexts::Chunked {
-        request_id,
-        chunk_id,
-        start,
-        chunks,
-        end,
-    }
-}
-
-#[cfg(test)]
-fn byte_stream_frame_texts(byte_stream: ControlByteStream, chunk_bytes: usize) -> Vec<String> {
-    match byte_stream_frame_text_parts(byte_stream, chunk_bytes) {
-        ControlFrameTexts::Immediate(frames) => frames,
-        ControlFrameTexts::Chunked {
-            start, chunks, end, ..
-        } => {
-            let mut frames = Vec::with_capacity(chunks.len() + 2);
-            frames.push(start);
-            frames.extend(chunks);
-            frames.push(end);
-            frames
-        }
-    }
-}
-
-fn control_frame_text_parts(
-    frame: serde_json::Value,
-    threshold_bytes: usize,
-    chunk_bytes: usize,
-) -> ControlFrameTexts {
-    let text = frame.to_string();
-    let frame_type = frame.get("t").and_then(|v| v.as_str());
-    if !matches!(frame_type, Some("response") | Some("stream_event"))
-        || text.len() <= threshold_bytes
-        || chunk_bytes == 0
-    {
-        return ControlFrameTexts::Immediate(vec![text]);
-    }
-    let request_id = frame
-        .get("id")
-        .and_then(|v| v.as_str())
-        .unwrap_or("")
-        .to_string();
-    if request_id.is_empty() {
-        return ControlFrameTexts::Immediate(vec![text]);
-    }
-    let chunk_id = frame
-        .get("chunk_id")
-        .and_then(|v| v.as_str())
-        .filter(|v| !v.is_empty())
-        .map(ToString::to_string)
-        .or_else(|| {
-            frame
-                .get("seq")
-                .and_then(|v| v.as_u64())
-                .map(|seq| format!("{request_id}:{seq}"))
-        })
-        .unwrap_or_else(|| request_id.clone());
-
-    let bytes = text.as_bytes();
-    let chunk_count = bytes.len().div_ceil(chunk_bytes);
-    let start = serde_json::json!({
-        "t": "response_start",
-        "id": request_id,
-        "chunk_id": chunk_id,
-        "encoding": "base64-json-frame",
-        "total_bytes": bytes.len(),
-        "chunks": chunk_count,
-    })
-    .to_string();
-    let mut chunks = Vec::with_capacity(chunk_count);
-    for (seq, chunk) in bytes.chunks(chunk_bytes).enumerate() {
-        chunks.push(
-            serde_json::json!({
-                "t": "response_chunk",
-                "id": request_id,
-                "chunk_id": chunk_id,
-                "seq": seq,
-                "data": base64::engine::general_purpose::STANDARD.encode(chunk),
-            })
-            .to_string(),
-        );
-    }
-    let end = serde_json::json!({
-        "t": "response_end",
-        "id": request_id,
-        "chunk_id": chunk_id,
-        "chunks": chunk_count,
-    })
-    .to_string();
-    ControlFrameTexts::Chunked {
-        request_id,
-        chunk_id,
-        start,
-        chunks,
-        end,
-    }
-}
-
-#[cfg(test)]
-fn chunk_control_response_frame(
-    frame: serde_json::Value,
-    threshold_bytes: usize,
-    chunk_bytes: usize,
-) -> Vec<String> {
-    match control_frame_text_parts(frame, threshold_bytes, chunk_bytes) {
-        ControlFrameTexts::Immediate(frames) => frames,
-        ControlFrameTexts::Chunked {
-            start, chunks, end, ..
-        } => {
-            let mut frames = Vec::with_capacity(chunks.len() + 2);
-            frames.push(start);
-            frames.extend(chunks);
-            frames.push(end);
-            frames
-        }
-    }
-}
-
-fn drain_queued_control_frames<I: rtc::interceptor::Interceptor>(
-    rtc: &mut RTCPeerConnection<I>,
-    channels: &HashMap<String, rtc::data_channel::RTCDataChannelId>,
-    outbound_queue: &mut OutboundControlQueue,
-) {
-    loop {
-        let mut pop_front = false;
-        let mut completed_end: Option<String> = None;
-        match outbound_queue.frames.front_mut() {
-            Some(QueuedControlFrame::Immediate { text, .. }) => {
-                send_control_text(rtc, channels, text.clone());
-                pop_front = true;
-            }
-            Some(QueuedControlFrame::Chunked(queued)) => {
-                if !queued.started {
-                    send_control_text(rtc, channels, queued.start.clone());
-                    queued.started = true;
-                }
-                while queued.credit > 0 && queued.next_chunk < queued.chunks.len() {
-                    let text = queued.chunks[queued.next_chunk].clone();
-                    queued.next_chunk += 1;
-                    queued.credit -= 1;
-                    send_control_text(rtc, channels, text);
-                }
-                if queued.next_chunk >= queued.chunks.len() {
-                    completed_end = Some(queued.end.clone());
-                }
-            }
-            None => break,
-        }
-        if let Some(end) = completed_end {
-            outbound_queue.frames.pop_front();
-            send_control_text(rtc, channels, end);
-            continue;
-        }
-        if pop_front {
-            outbound_queue.frames.pop_front();
-            continue;
-        }
-        break;
-    }
-}
-
-fn dashboard_control_error_response(id: String, message: impl Into<String>) -> serde_json::Value {
-    serde_json::json!({
-        "t": "response",
-        "id": id,
-        "ok": false,
-        "error": message.into(),
-    })
 }
 
 fn runtime_allows_operation(
@@ -14306,168 +13452,4 @@ mod tests {
         assert_eq!(frame["result"]["_httpOk"], false);
     }
 
-    #[test]
-    fn oversized_response_frames_are_chunked_and_reassemble() {
-        let frame = serde_json::json!({
-            "t": "response",
-            "id": "large-1",
-            "ok": true,
-            "result": {
-                "text": "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-            }
-        });
-        let original = frame.to_string();
-        let frames = chunk_control_response_frame(frame, 40, 12);
-        assert!(frames.len() > 3, "expected start/chunks/end frames");
-
-        let start: serde_json::Value = serde_json::from_str(&frames[0]).unwrap();
-        assert_eq!(start["t"], "response_start");
-        assert_eq!(start["id"], "large-1");
-        assert_eq!(start["encoding"], "base64-json-frame");
-        assert_eq!(start["total_bytes"], original.len());
-
-        let end: serde_json::Value = serde_json::from_str(frames.last().unwrap()).unwrap();
-        assert_eq!(end["t"], "response_end");
-        assert_eq!(end["id"], "large-1");
-        assert_eq!(end["chunks"], start["chunks"]);
-
-        let mut bytes = Vec::new();
-        for (seq, text) in frames[1..frames.len() - 1].iter().enumerate() {
-            let chunk: serde_json::Value = serde_json::from_str(text).unwrap();
-            assert_eq!(chunk["t"], "response_chunk");
-            assert_eq!(chunk["id"], "large-1");
-            assert_eq!(chunk["seq"], seq);
-            let encoded = chunk["data"].as_str().unwrap();
-            bytes.extend(
-                base64::engine::general_purpose::STANDARD
-                    .decode(encoded)
-                    .unwrap(),
-            );
-        }
-        assert_eq!(String::from_utf8(bytes).unwrap(), original);
-    }
-
-    #[test]
-    fn oversized_stream_event_frames_are_chunked_with_chunk_ids() {
-        let frame = serde_json::json!({
-            "t": "stream_event",
-            "id": "stream-1",
-            "seq": 7,
-            "chunk_id": "stream-1:7",
-            "event": {
-                "type": "replace",
-                "sessions": ["x".repeat(128)]
-            }
-        });
-        let frames = chunk_control_response_frame(frame.clone(), 40, 24);
-        assert!(frames.len() > 3, "expected stream event chunking");
-
-        let start: serde_json::Value = serde_json::from_str(&frames[0]).unwrap();
-        assert_eq!(start["t"], "response_start");
-        assert_eq!(start["id"], "stream-1");
-        assert_eq!(start["chunk_id"], "stream-1:7");
-
-        let end: serde_json::Value = serde_json::from_str(frames.last().unwrap()).unwrap();
-        assert_eq!(end["t"], "response_end");
-        assert_eq!(end["id"], "stream-1");
-        assert_eq!(end["chunk_id"], "stream-1:7");
-
-        let mut bytes = Vec::new();
-        for text in &frames[1..frames.len() - 1] {
-            let chunk: serde_json::Value = serde_json::from_str(text).unwrap();
-            assert_eq!(chunk["t"], "response_chunk");
-            assert_eq!(chunk["id"], "stream-1");
-            assert_eq!(chunk["chunk_id"], "stream-1:7");
-            bytes.extend(
-                base64::engine::general_purpose::STANDARD
-                    .decode(chunk["data"].as_str().unwrap())
-                    .unwrap(),
-            );
-        }
-        assert_eq!(String::from_utf8(bytes).unwrap(), frame.to_string());
-    }
-
-    #[test]
-    fn byte_stream_frames_are_chunked_and_credit_addressable() {
-        let bytes: Vec<u8> = (0..73).map(|i| (i % 251) as u8).collect();
-        let stream = ControlByteStream {
-            id: "download-1".to_string(),
-            stream_id: "download-1:file".to_string(),
-            content_type: "application/octet-stream".to_string(),
-            filename: Some("artifact.bin".to_string()),
-            bytes: bytes.clone(),
-            result: serde_json::json!({
-                "ok": true,
-                "filename": "artifact.bin",
-                "size": bytes.len(),
-            }),
-        };
-        let frames = byte_stream_frame_texts(stream, 13);
-        assert_eq!(frames.len(), 8, "expected start + 6 chunks + end");
-
-        let start: serde_json::Value = serde_json::from_str(&frames[0]).unwrap();
-        assert_eq!(start["t"], "byte_stream_start");
-        assert_eq!(start["id"], "download-1");
-        assert_eq!(start["stream_id"], "download-1:file");
-        assert_eq!(start["encoding"], "base64");
-        assert_eq!(start["content_type"], "application/octet-stream");
-        assert_eq!(start["filename"], "artifact.bin");
-        assert_eq!(start["total_bytes"], bytes.len());
-        assert_eq!(start["chunks"], 6);
-
-        let mut decoded = Vec::new();
-        for (seq, text) in frames[1..frames.len() - 1].iter().enumerate() {
-            let chunk: serde_json::Value = serde_json::from_str(text).unwrap();
-            assert_eq!(chunk["t"], "byte_stream_chunk");
-            assert_eq!(chunk["id"], "download-1");
-            assert_eq!(chunk["stream_id"], "download-1:file");
-            assert_eq!(chunk["seq"], seq);
-            decoded.extend(
-                base64::engine::general_purpose::STANDARD
-                    .decode(chunk["data"].as_str().unwrap())
-                    .unwrap(),
-            );
-        }
-        assert_eq!(decoded, bytes);
-
-        let end: serde_json::Value = serde_json::from_str(frames.last().unwrap()).unwrap();
-        assert_eq!(end["t"], "byte_stream_end");
-        assert_eq!(end["id"], "download-1");
-        assert_eq!(end["stream_id"], "download-1:file");
-        assert_eq!(end["chunks"], 6);
-        assert_eq!(end["result"]["filename"], "artifact.bin");
-    }
-
-    #[test]
-    fn default_response_chunks_stay_below_datachannel_edge() {
-        let frame = serde_json::json!({
-            "t": "response",
-            "id": "large-2",
-            "ok": true,
-            "result": {
-                "text": "x".repeat(CONTROL_RESPONSE_CHUNK_THRESHOLD_BYTES * 4)
-            }
-        });
-        let frames = control_frame_texts(frame);
-        assert!(frames.len() > 3, "expected default chunking");
-
-        for text in frames {
-            let parsed: serde_json::Value = serde_json::from_str(&text).unwrap();
-            if parsed["t"] == "response_chunk" {
-                assert!(
-                    text.len() < 32 * 1024,
-                    "chunk frame is too close to common DataChannel limits: {} bytes",
-                    text.len()
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn small_or_non_response_frames_are_not_chunked() {
-        let response = serde_json::json!({"t":"response","id":"small","ok":true,"result":{}});
-        assert_eq!(chunk_control_response_frame(response, 4096, 16).len(), 1);
-        let event = serde_json::json!({"t":"event","id":"e1","payload":{"text":"large enough"}});
-        assert_eq!(chunk_control_response_frame(event, 1, 1).len(), 1);
-    }
 }
